+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.10.1
+* Match version declaration of vertex shader
+
+## 2.10.0
+* Try to acquire a GLES 3.0 context if available
+
 ## 2.9.1
 * Fix broken backbuffer preset selection
 * Fix crash when setting backbuffer parameters